requirements.txt
# Core Flask Framework
Flask==3.0.3
Werkzeug==3.0.6

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0

# Forms and Validation
WTForms==3.1.2

# Database
SQLAlchemy==2.0.36
Alembic==1.13.2

# Security and Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
itsdangerous==2.2.0

# Configuration and Environment
python-dotenv==1.0.1

# CLI and Development
click==8.1.7

# Production Server
gunicorn==23.0.0

# Development Tools
pytest==8.3.4
pytest-flask==1.3.0
coverage==7.6.1

# Performance Monitoring
psutil==6.0.0

# Timezone handling
pytz==2025.2

# Additional utilities
requests==2.32.3
Jinja2==3.1.4
MarkupSafe==2.1.5

# Push notifications / JWT
cryptography==43.0.1
PyJWT==2.9.0

# Python compatibility
setuptools>=65.0.0

# Project-specific
twilio==9.3.7
anthropic==0.40.0
APScheduler==3.10.4
xhtml2pdf==0.2.16
openpyxl==3.1.5
